Add --enable-this-cpu option.
Especially on Windows, all builds tend to be cross-compiling builds, but where
the host machine is the same as the build machine. This means it is not
possible to have CPU-specific optimisations. The --enable-this-cpu configure
option (disabled by default) does a little bit of trickery to allow the normal
determining of the CPU architecture to run, but still main the generic cross-
compiling tool prefixes (e.g. still use x86_64-x64-mingw32-gcc but have
Broadwell-specific optimisations).

Signed-off-by: David Allsopp <David.Allsopp@cl.cam.ac.uk>